refactor: move `custom_math` to `core`
